Consolidate offset tests into one file
